Time: 0 ms (100%), Space: 7.1 MB (49.03%) - LeetHub

+++ 0231-power-of-two/0231-power-of-two.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
 
-    bool power(float n) {
+    bool power(double n) {
         if(n == 1 ) {
             return true;
         }
